+++ src/main/java/mju/scholarship/member/dto/UpdateMemberInfoRequest.java
@@ -18,6 +18,6 @@ public class UpdateMemberInfoRequest {
     private String gender; //enum 으로 할까
     private String city;
     private String department; // major -> 전공 (?)
-    private Integer grade; // 1.0 - 4.5
+    private Double grade; // 1.0 - 4.5
     private Integer incomeQuantile; // 1-10
 }
