+++ src/main/java/mju/scholarship/scholoarship/dto/CreateScholarshipRequest.java
@@ -15,6 +15,6 @@ public class CreateScholarshipRequest {
     private String province;
     private String city;
     private String department; // major -> 전공 (?)
-    private Integer grade; // 1.0 - 4.5
+    private Double grade; // 1.0 - 4.5
     private Integer incomeQuantile; // 1-10
 }
